Consolidate and use asCompatibleWithIfPossible
Previously:
- PackageRange.withTerseConstraint converted constraints to strings to check
  if they started with '^' (constraint.toString().startsWith('^')).
- Term constructor invoked withTerseConstraint on every term creation.
- A redundant copy of _compatibleWithIfPossible existed in dependency_services.dart.
- sdk.dart defined an AsCompatibleWithIfPossible extension.

This change:
- Moves AsCompatibleWithIfPossible to utils.dart (re-exported by sdk.dart).
- Replaces duplicate helpers with asCompatibleWithIfPossible across pub.
- Formats constraints in PackageRange.toString using asCompatibleWithIfPossible.
- Simplifies Term constructor to Term(this.package, this.isPositive).

TAG=agy
CONV=06fb411e-8696-4209-979a-5b0d67624eb4
