+++ lib/src/sdk.dart
@@ -9,6 +9,7 @@ import 'package:pub_semver/pub_semver.dart';
 import 'sdk/dart.dart';
 import 'sdk/flutter.dart';
 import 'sdk/fuchsia.dart';
+export 'utils.dart' show AsCompatibleWithIfPossible;
 
 /// An SDK that can provide packages and on which pubspecs can express version
 /// constraints.
@@ -61,16 +62,3 @@ final sdks = UnmodifiableMapView<String, Sdk>({
 
 /// The core Dart SDK.
 final sdk = DartSdk();
-
-extension AsCompatibleWithIfPossible on VersionConstraint {
-  // Returns `this` expressed as [VersionConstraint.compatibleWith] if possible.
-  VersionConstraint asCompatibleWithIfPossible() {
-    final range = this;
-    if (range is! VersionRange) return this;
-    final min = range.min;
-    if (min == null) return this;
-    final asCompatibleWith = VersionConstraint.compatibleWith(min);
-    if (asCompatibleWith == this) return asCompatibleWith;
-    return this;
-  }
-}
