+++ lib/src/solver/incompatibility.dart
@@ -3,6 +3,7 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import '../package_name.dart';
+import '../utils.dart';
 import 'incompatibility_cause.dart';
 import 'term.dart';
 
@@ -131,7 +132,7 @@ class Incompatibility {
       assert(terms.length == 1);
       assert(terms.first.isPositive);
       return 'no versions of ${_terseRef(terms.first, details)} '
-          'match ${terms.first.constraint}';
+          'match ${terms.first.constraint.asCompatibleWithIfPossible()}';
     } else if (cause is PackageNotFoundIncompatibilityCause) {
       assert(terms.length == 1);
       assert(terms.first.isPositive);
