+++ lib/src/utils.dart
@@ -838,3 +838,17 @@ String readableFileSize(int size) {
     return '<1 KB';
   }
 }
+
+extension AsCompatibleWithIfPossible on VersionConstraint {
+  /// Returns `this` expressed as [VersionConstraint.compatibleWith] if
+  /// possible.
+  VersionConstraint asCompatibleWithIfPossible() {
+    final range = this;
+    if (range is! VersionRange) return this;
+    final min = range.min;
+    if (min == null) return this;
+    final asCompatibleWith = VersionConstraint.compatibleWith(min);
+    if (asCompatibleWith == this) return asCompatibleWith;
+    return this;
+  }
+}
